fix configuration and tests to be able to run tests again after RabbitMQ configuration
