+++ src/main/java/com/jobfinder/finder/config/rabbitMq/RabbitMQConfiguration.java
@@ -5,8 +5,10 @@ import org.springframework.amqp.core.BindingBuilder;
 import org.springframework.amqp.core.Queue;
 import org.springframework.amqp.core.TopicExchange;
 import org.springframework.beans.factory.annotation.Value;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.context.annotation.Bean;
 
+@ConditionalOnProperty("${spring.rabbitmq.enabled:false}")
 public class RabbitMQConfiguration {
 
   @Value("${spring.rabbitmq.host:localhost}")
