+++ src/main/java/com/jobfinder/finder/service/EmailService.java
@@ -1,6 +1,7 @@
 package com.jobfinder.finder.service;
 
 import lombok.RequiredArgsConstructor;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.mail.SimpleMailMessage;
 import org.springframework.mail.javamail.JavaMailSender;
 import org.springframework.stereotype.Service;
