+++ src/test/java/com/jobfinder/finder/integrationTest/controller/FinderIntegrationTestInitiator.java
@@ -2,13 +2,11 @@ package com.jobfinder.finder.integrationTest.controller;
 
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.jobfinder.finder.controller.UserController;
-import com.jobfinder.finder.integrationTest.configuration.MockUserDetailsManagerConfig;
+import com.jobfinder.finder.integrationTest.configuration.JobFinderTestConfiguration;
 import org.junit.jupiter.api.BeforeEach;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
 import org.springframework.boot.autoconfigure.data.redis.RedisAutoConfiguration;
-import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.context.annotation.Import;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.web.servlet.MockMvc;
@@ -17,12 +15,13 @@ import org.springframework.web.context.WebApplicationContext;
 
 import static org.springframework.security.test.web.servlet.setup.SecurityMockMvcConfigurers.springSecurity;
 
-@Import(MockUserDetailsManagerConfig.class)
 @EnableAutoConfiguration(exclude = {
     org.springframework.boot.autoconfigure.jdbc.DataSourceAutoConfiguration.class,
     org.springframework.boot.autoconfigure.orm.jpa.HibernateJpaAutoConfiguration.class,
-    RedisAutoConfiguration.class
+    RedisAutoConfiguration.class,
+    org.springframework.boot.autoconfigure.amqp.RabbitAutoConfiguration.class
 })
+@Import(JobFinderTestConfiguration.class)
 @ActiveProfiles("test")
 public abstract class FinderIntegrationTestInitiator {
   //this class is used to initiate integration tests for the Finder application
