+++ src/test/java/com/jobfinder/finder/integrationTest/service/CacheTestIntializer.java
@@ -1,7 +1,8 @@
 package com.jobfinder.finder.integrationTest.service;
 
 import com.jobfinder.finder.config.redis.RedisConfiguration;
-import com.jobfinder.finder.integrationTest.configuration.CacheTestConfiguration;
+import com.jobfinder.finder.integrationTest.configuration.MockUserDetailsManagerConfig;
+import com.jobfinder.finder.integrationTest.configuration.JobFinderTestConfiguration;
 import java.util.Objects;
 import org.junit.jupiter.api.BeforeEach;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -9,7 +10,7 @@ import org.springframework.cache.CacheManager;
 import org.springframework.cache.annotation.EnableCaching;
 import org.springframework.context.annotation.Import;
 
-@Import(CacheTestConfiguration.class)
+@Import({JobFinderTestConfiguration.class})
 @EnableCaching
 public abstract class CacheTestIntializer {
